refactor(footer): decompose Footer into focused sub-components
Footer failed the "and" test: it rendered nav links, locale picker,
copyright, fineprint, and image link in one 300-line file. Split into:

- Footer.types.ts  — exported interfaces (own reason to change)
- FooterNavLinks.tsx — <nav> + link list + FooterAnchor
- FooterLocalePicker.tsx — NativeSelect or Skeleton; early-return branch
  replaces the isLoading ternary
- Footer.tsx — thin orchestrator (~80 lines); FooterImageAnchor stays
  here as it is tightly coupled to the root's extLinkDescId
