+++ frontend/packages/component-library/src/footer/Footer.tsx
@@ -13,150 +13,76 @@
  * @see {@link FooterProps}
  */
 
-import {
-  Box,
-  FormControl,
-  Grid,
-  InputLabel,
-  Link,
-  List,
-  ListItem,
-  NativeSelect,
-  Skeleton,
-  Typography,
-} from '@mui/material';
+import {Box, Grid, Link, Typography} from '@mui/material';
 import type {LinkProps} from '@mui/material/Link';
 import {styled} from '@mui/material/styles';
 import {visuallyHidden} from '@mui/utils';
 import {useId} from 'react';
-import type {ReactNode} from 'react';
 
-// ---------------------------------------------------------------------------
-// Prop interfaces
-// ---------------------------------------------------------------------------
+import type {FooterImageLink, FooterProps} from './Footer.types';
+import FooterLocalePicker from './FooterLocalePicker';
+import FooterNavLinks from './FooterNavLinks';
 
-/** A site navigation link rendered in the footer link bar. */
-export interface FooterSiteLink {
-  id: string;
-  label: string;
-  href: string;
-  /** When true, opens in a new tab with rel="noopener noreferrer". */
-  external?: boolean;
-  /** When true the link renders with the theme's accent color (e.g. brand orange). */
-  accent?: boolean;
-}
-
-/** The attribution image link rendered at the bottom. */
-export interface FooterImageLink {
-  src: string;
-  altText: string;
-  href: string;
-  /** When true, opens in a new tab with rel="noopener noreferrer". */
-  external?: boolean;
-}
-
-/** A single option in the language picker. */
-export interface FooterLanguageOption {
-  /** Language code passed to onLanguageChange. */
-  value: string;
-  /** Human-readable language name rendered in the option. */
-  text: string;
-}
-
-/** Props for the Footer design-system primitive. */
-export interface FooterProps {
-  /** Navigation links rendered as a pipe-separated inline list. */
-  siteLinks: FooterSiteLink[];
-  /**
-   * Pre-composed copyright / trademark line. Typically a ReactNode produced
-   * by the studio composer so the year `<span>` selector stays at the source.
-   */
-  copyright: ReactNode;
-  /** Art-credits + trademark block, rendered below copyright. */
-  fineprint?: ReactNode;
-  /** Attribution image + link, rendered at the very bottom. */
-  imageLink?: FooterImageLink;
-  /**
-   * Available languages for the locale picker, or `'loading'` to show a
-   * skeleton while the list is being fetched. Passing `'loading'` avoids a
-   * 1→N option flicker when the language list resolves asynchronously.
-   */
-  languages: FooterLanguageOption[] | 'loading';
-  /** Currently selected locale code (controls the picker's selected option). */
-  selectedLocaleCode: string;
-  /** Called with the chosen language code when the user changes the picker. */
-  onLanguageChange: (code: string) => void;
-  className?: string;
-}
+export type {
+  FooterSiteLink,
+  FooterImageLink,
+  FooterLanguageOption,
+  FooterProps,
+} from './Footer.types';
 
 // ---------------------------------------------------------------------------
 // Slot wrappers — empty bodies; all paint lives in the theme override
 // ---------------------------------------------------------------------------
 
-const FooterRoot = styled('footer', {
-  name: 'MuiFooter',
-  slot: 'root',
-})({});
-
-const FooterGrid = styled(Grid, {
-  name: 'MuiFooter',
-  slot: 'grid',
-})({});
-
-const FooterLinks = styled(List, {
-  name: 'MuiFooter',
-  slot: 'links',
-})({});
-
-const FooterLinkItem = styled('span', {
-  name: 'MuiFooter',
-  slot: 'link',
-})({});
-
-const FooterLocaleSelect = styled('div', {
-  name: 'MuiFooter',
-  slot: 'localeSelect',
-})({});
-
+const FooterRoot = styled('footer', {name: 'MuiFooter', slot: 'root'})({});
+const FooterGrid = styled(Grid, {name: 'MuiFooter', slot: 'grid'})({});
 const FooterCopyright = styled(Typography, {
   name: 'MuiFooter',
   slot: 'copyright',
 })({});
-
 const FooterFineprint = styled(Typography, {
   name: 'MuiFooter',
   slot: 'fineprint',
 })({});
-
-const ImageLinkWrapper = styled('div', {
-  name: 'MuiFooter',
-  slot: 'imageLink',
-})({});
+const ImageLinkWrapper = styled('div', {name: 'MuiFooter', slot: 'imageLink'})(
+  {},
+);
 
 // ---------------------------------------------------------------------------
-// Private FooterAnchor wrapper
+// ImageLink — local anchor with external-link semantics
 // ---------------------------------------------------------------------------
 
 // TODO(footer): remove once DSCO `link` → MUI `Link` migration lands
 // (per MIGRATION_STATUS.md#link) and MUI `Link` carries `external` natively.
-const FooterAnchor = ({
-  external,
-  children,
-  ...props
-}: LinkProps & {external?: boolean}) => (
-  <Link
-    {...props}
-    {...(external ? {rel: 'noopener noreferrer', target: '_blank'} : {})}
-  >
-    {children}
-  </Link>
+const buildImageLinkProps = (external?: boolean) =>
+  external ? {rel: 'noopener noreferrer', target: '_blank'} : {};
+
+interface FooterImageAnchorProps extends Omit<LinkProps, 'href'> {
+  imageLink: FooterImageLink;
+  extLinkDescId: string;
+}
+
+/** Attribution image wrapped in a themed anchor. */
+const FooterImageAnchor = ({
+  imageLink,
+  extLinkDescId,
+  ...rest
+}: FooterImageAnchorProps) => (
+  <ImageLinkWrapper>
+    <Link
+      href={imageLink.href}
+      aria-describedby={imageLink.external ? extLinkDescId : undefined}
+      {...buildImageLinkProps(imageLink.external)}
+      {...rest}
+    >
+      <img src={imageLink.src} alt={imageLink.altText} loading="lazy" />
+    </Link>
+  </ImageLinkWrapper>
 );
 
 // ---------------------------------------------------------------------------
-// Footer component
+// Footer
 // ---------------------------------------------------------------------------
-
-const LOCALE_SELECT_ID = 'footer-locale-select';
 
 /**
  * Site footer with navigation links, language picker, copyright, fineprint,
@@ -174,7 +100,6 @@ const Footer = ({
   onLanguageChange,
   className,
 }: FooterProps) => {
-  const isLoading = languages === 'loading';
   const extLinkDescId = useId();
 
   return (
@@ -188,7 +113,6 @@ const Footer = ({
         (opens in a new tab)
       </Box>
       <FooterGrid container>
-        {/* Top: links (left) and locale picker (right), picker wraps above links on mobile */}
         <Grid
           size={12}
           display="flex"
@@ -197,60 +121,18 @@ const Footer = ({
           gap={2}
         >
           <Grid size={{md: 'grow'}}>
-            <nav aria-label="Footer">
-              <FooterLinks disablePadding>
-                {siteLinks.map(link => (
-                  <ListItem key={link.id} disablePadding disableGutters>
-                    <FooterLinkItem>
-                      <FooterAnchor
-                        href={link.href}
-                        external={link.external}
-                        data-accent={link.accent || undefined}
-                        aria-describedby={
-                          link.external ? extLinkDescId : undefined
-                        }
-                      >
-                        {link.label}
-                      </FooterAnchor>
-                    </FooterLinkItem>
-                  </ListItem>
-                ))}
-              </FooterLinks>
-            </nav>
+            <FooterNavLinks
+              siteLinks={siteLinks}
+              extLinkDescId={extLinkDescId}
+            />
           </Grid>
-          {isLoading ? (
-            <FooterLocaleSelect>
-              <Skeleton
-                variant="rectangular"
-                width="8.5rem"
-                height="1.5rem"
-                role="status"
-                aria-label="Loading language options"
-              />
-            </FooterLocaleSelect>
-          ) : (
-            <FooterLocaleSelect>
-              <FormControl>
-                <InputLabel htmlFor={LOCALE_SELECT_ID} sx={visuallyHidden}>
-                  Language
-                </InputLabel>
-                <NativeSelect
-                  value={selectedLocaleCode}
-                  onChange={e => onLanguageChange(e.target.value)}
-                  inputProps={{id: LOCALE_SELECT_ID}}
-                >
-                  {languages.map(lang => (
-                    <option key={lang.value} value={lang.value}>
-                      {lang.text}
-                    </option>
-                  ))}
-                </NativeSelect>
-              </FormControl>
-            </FooterLocaleSelect>
-          )}
+          <FooterLocalePicker
+            languages={languages}
+            selectedLocaleCode={selectedLocaleCode}
+            onLanguageChange={onLanguageChange}
+          />
         </Grid>
 
-        {/* Copyright — right-aligned to sit below the locale picker */}
         <Grid size={12}>
           <FooterCopyright
             component="p"
@@ -261,7 +143,6 @@ const Footer = ({
           </FooterCopyright>
         </Grid>
 
-        {/* Fineprint */}
         {fineprint && (
           <Grid size={12}>
             <FooterFineprint
@@ -274,24 +155,12 @@ const Footer = ({
           </Grid>
         )}
 
-        {/* Attribution image */}
         {imageLink && (
           <Grid size={12}>
-            <ImageLinkWrapper>
-              <FooterAnchor
-                href={imageLink.href}
-                external={imageLink.external}
-                aria-describedby={
-                  imageLink.external ? extLinkDescId : undefined
-                }
-              >
-                <img
-                  src={imageLink.src}
-                  alt={imageLink.altText}
-                  loading="lazy"
-                />
-              </FooterAnchor>
-            </ImageLinkWrapper>
+            <FooterImageAnchor
+              imageLink={imageLink}
+              extLinkDescId={extLinkDescId}
+            />
           </Grid>
         )}
       </FooterGrid>
